- updated to latest freicore, which brings in Boost 1.56, newer pwiz, and Visual Studio 2013 support
Former-commit-id: 0bc4ae6dd1dcc657acec394f94ec0f536b9a5d2f

+++ pwiz_tools/Bumbershoot/directag/directagSpectrum.h
@@ -89,7 +89,7 @@ namespace directag
 	{
 		Spectrum();
 		Spectrum( const Spectrum& old );
-        Spectrum( const string& nativeID, size_t charge, double precursorMZ, const flat_map<double,double>& peaks);
+        Spectrum( const string& nativeID, size_t charge, double precursorMZ, const flat_map<double, float>& peaks);
 
 		void initialize( int numIntenClasses, int NumMzFidelityClasses )
 		{
